add deploy extensions script
