+++ src/deploy/deploy_extensions.ts
@@ -1,0 +1,77 @@
+import { DeployFunction } from "hardhat-deploy/types";
+import { HardhatRuntimeEnvironment } from "hardhat/types";
+
+const deploy: DeployFunction = async function (
+  hre: HardhatRuntimeEnvironment,
+) {
+  const { deployments, getNamedAccounts } = hre;
+  const { deployer } = await getNamedAccounts();
+  const { deploy } = deployments;
+
+  const authority = await deploy("Authority", {
+    from: deployer,
+    args: [deployer],
+    log: true,
+    deterministicDeployment: true,
+  });
+
+  const registry = await deploy("PoolRegistry", {
+    from: deployer,
+    args: [
+      authority.address,
+      deployer  // Rigoblock Dao
+    ],
+    log: true,
+    deterministicDeployment: true,
+  });
+
+  const poolImplementation = await deploy("RigoblockV3Pool", {
+    from: deployer,
+    args: [authority.address],
+    log: true,
+    deterministicDeployment: true,
+  });
+
+  const factory = await deploy("RigoblockPoolProxyFactory", {
+    from: deployer,
+    args: [
+      poolImplementation.address,
+      registry.address
+    ],
+    log: true,
+    deterministicDeployment: true,
+  });
+
+  await deploy("EWhitelist", {
+    from: deployer,
+    args: [authority.address],
+    log: true,
+    deterministicDeployment: true,
+  });
+
+  await deploy("EUpgrade", {
+    from: deployer,
+    args: [factory.address],
+    log: true,
+    deterministicDeployment: true,
+  });
+
+  /*const uniswapRouter2 = "0x68b3465833fb72A70ecDF485E0e4C7bD8665Fc45"
+
+  await deploy("AUniswap", {
+    from: deployer,
+    args: [uniswapRouter2],
+    log: true,
+    deterministicDeployment: true,
+  });*/
+
+  await deploy("AMulticall", {
+    from: deployer,
+    args: [],
+    log: true,
+    deterministicDeployment: true,
+  });
+};
+
+deploy.tags = ['extensions', 'adapters', 'l2-suite', 'main-suite']
+export default deploy;
